requirements.txt
Keras==2.4.3
Keras-Preprocessing==1.1.2
matplotlib==3.5.3
numexpr==2.8.4
numpy==1.19.5
PyWavelets==1.3.0
scipy==1.7.3
tables==3.7.0
tensorflow-gpu==2.4.0

